fix bug in unassign and reorganize code
