+++ basics/src/pages/configurator/index.css
@@ -1,5 +1,4 @@
 .configurator {
   width: 100%;
-  margin: 0 70px;
   color: rgb(2, 34, 90);
 }
